Fix exception: NoMethodError: "undefined method `status' for #<Faraday::TimeoutError:...>"
Somehow Faraday::TimeoutError object is passed as `response` parameter. In this case it doesn't have status or body, but the request url/headers still matter.
Probably the bug is in a different place that passes the invalid `response` parameter, but even if that bug will be found and fixed, this change can stay as a failsafe.